% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimorph_0.1_functions_bootstrap_and_specific_resampling_procedures_2025-06-07.r
\name{bootdimorph}
\alias{bootdimorph}
\title{Bootstrap dimorphism confidence intervals in univariate or multivariate sample}
\usage{
bootdimorph(
  x,
  sex = NULL,
  methsUni = c("SSD", "MMR", "BDI", "ERM", "FMA", "MoM", "BFM", "CV", "CVsex", "sdlog",
    "sdlogsex"),
  methsMulti = "GMM",
  nResamp = 1000,
  exact = F,
  limit = 50000,
  center = "geomean",
  sex.female = 1,
  na.rm = T,
  ncorrection = F,
  struc = NULL,
  datastruc = NULL,
  templatevar = NULL,
  alternative = "two.sided",
  conf.level = 0.95
)
}
\arguments{
\item{x}{A matrix or data frame of measurements from a comparative sample, with rows
corresponding to individual specimens and columns corresponding to size variables.  Sex data
should not be included.}

\item{sex}{A vector indicating sex for the individuals in \code{comparative}.  Defaults to \code{NULL}.}

\item{methsUni}{A character vector specifying the univariate method(s) used to calculate or estimate dimorphism. 
See \code{\link{dimorph}} for options.}

\item{methsMulti}{A character vector specifying the multivariate method(s) used to 
calculate or estimate dimorphism.  Note that regardless of the value of this argument,
multivariate estimation procedures will only be carried out if \code{x} is a multivariate 
dataset.  See \code{\link{dimorph}} for options.}

\item{nResamp}{Integer specifying the number of resampling iterations if Monte Carlo sampling is used.}

\item{exact}{Logical scalar specifying whether to sample all unique combinations of resampled datasets at the same
sample size as \code{x} from \code{x} when sampling with replacement.  Defaults to \code{FALSE}.  If set to 
\code{FALSE}, or if set to \code{TRUE} and the number of unique combinations exceeds \code{limit}, then Monte  
Carlo sampling is used instead.}

\item{limit}{Integer setting the upper limit on the number of unique combinations allowable 
for exact resampling.  If exact resampling would produce more resampled datasets than this number, 
Monte Carlo resampling is used instead.  Defaults to 50,000.}

\item{center}{A character string specifying the method used to calculate a mean, either \code{"geomean"} 
(default) which uses the geometric mean, or \code{"mean"} which uses the arithmetic mean.  More broadly, 
\code{"geomean"} indicates analyses are conducted in logarithmic data space and \code{"mean"} indicates 
analyses are conducted in raw data space.  Some methods can only be applied in one domain or the other: 
\code{"CV"} and \code{"CVsex"} are always calculated in raw data space and \code{center} will be set to 
\code{"mean"} for these methods regardless of the value set by the user; \code{"MoM"}, \code{"sdlog"}, 
and \code{"sdlogsex"} are always calculated in logarithmic data space and \code{center} will be set to 
\code{"geomean"} for these methods regardless of the value set by the user.}

\item{sex.female}{An integer scalar (1 or 2) specifying which level of \code{sex} 
corresponds to female.  Ignored if \code{sex} is \code{NULL}.  Defaults to 1.}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be stripped before
the computation proceeds in univariate analyses.  Not relevant for multivariate analyses.  Defaults to \code{TRUE}.}

\item{ncorrection}{A logical scalar indicating whether to apply Sokal and Braumann's (1980) 
size correction factor to CV estimates.  Defaults to \code{FALSE}.}

\item{struc}{Not generally relevant for users, this argument is sometimes applicable when \code{bootdimorph}
is called by other functions.  Defaults to \code{NULL}.  See \code{\link{resampleSSD}} for more details.}

\item{datastruc}{Not generally relevant for users, this argument is sometimes applicable when \code{bootdimorph}
is called by other functions.  Defaults to \code{NULL}.  See \code{\link{resampleSSD}} for more details.}

\item{templatevar}{A character object or integer value specifying the name or column number 
of the variable in \code{x} to be estimated using the template method.  Ignored if template method 
is not used.  Defaults to \code{NULL}.}

\item{alternative}{A character object specifying whether to calculate two-sided (\code{"two.sided"}), or 
one-sided (\code{"less"} or \code{"greater"}) confidence intervals.  Defaults to \code{"two.sided"}.}

\item{conf.level}{Value between zero and one setting the confidence level, equal to 1-alpha. Defaults
to \code{0.95}.}
}
\value{
A list of class \code{dimorphResampledUni} or \code{dimorphResampledMulti} containing a dataframe
  with resampled dimorphism estimates and a \code{dimorphAds} object containg resampled addresses produced
  by \code{\link{getsampleaddresses}}.  Printing this object provides confidence intervals for all estimators calculated,
  and confidence intervals for bias from sample SSD if relevant.  Plotting this object produces violin plots for
  all bootstrapped distributions and lines indicating confidence limits.
}
\description{
Function to generate confidence intervals for actual or estimated dimorphism in a univariate or multivariate sample
  by percentile bootstrapping; i.e., sampling with replacement multiple times to a sample size equal to the original 
  dataset sample size, then discarding the highest and lowest (alpha/2) x 100 percent of resampled values to 
  find confidence limits (or discard the highest or the lowest alpha x 100 percent of values in the case of 
  one-sided confidence intervals).
}
\examples{
gor <- apelimbart[apelimbart$Species=="Gorilla gorilla",]
hom <- apelimbart[apelimbart$Species=="Homo sapiens",]
pan <- apelimbart[apelimbart$Species=="Pan troglodytes",]
hyl <- apelimbart[apelimbart$Species=="Hylobates lar",]
outcomeUgor <- bootdimorph(gor[,"FHSI", drop=FALSE], sex=gor$Sex, nResamp=100)
outcomeUhom <- bootdimorph(hom[,"FHSI", drop=FALSE], sex=hom$Sex, nResamp=100)
outcomeUpan <- bootdimorph(pan[,"FHSI", drop=FALSE], sex=pan$Sex, nResamp=100)
outcomeUhyl <- bootdimorph(hyl[,"FHSI", drop=FALSE], sex=hyl$Sex, nResamp=100)
outcomeUgor
plot(outcomeUgor)
outcomeUhom
plot(outcomeUhom)
outcomeUpan
plot(outcomeUpan)
outcomeUhyl
plot(outcomeUhyl)
plot(outcomeUgor, type="bias")
plot(outcomeUhom, type="bias")
plot(outcomeUpan, type="bias")
plot(outcomeUhyl, type="bias")
}
\seealso{
\code{\link{resampleSSD}}, \code{\link{dimorph}}, \code{\link{getsampleaddresses}}
}
