% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimorph_0.1_functions_general_2024-11-02.r
\name{centers}
\alias{centers}
\title{Central Tendencies for Resampled Distributions in \code{SSDtest} Object}
\usage{
centers(x)
}
\arguments{
\item{x}{An object of class \code{SSDtest}.}
}
\value{
A matrix containing the means and medians of resampled distributions in \code{x}.
}
\description{
Function for extracting the means and medians of resampled values from a \code{SSDtest} object as a matrix.
}
\examples{
# Standard significance test with one fossil sample, sampling without replacement
test_faux_uni <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI"]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI"],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI"],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "FHSI"],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI"]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("MMR", "BDI"),
     limit=1000,
     nResamp=100)
centers(test_faux_uni)
}
