% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimorph-package.R
\docType{package}
\name{dimorph-package}
\alias{dimorph-package}
\title{dimorph: Dimorphism Calculation and Estimation}
\description{
A system for estimating size dimorphism based on metric data using multiple published estimation techniques for univariate data and incomplete multivariate data, storing information related to the calculation of those estimates, and performing resampling-based significance tests for pairwise comparison of dimorphism estimates.
}
\author{
\strong{Maintainer}: Adam Gordon \email{agordon@albany.edu} (\href{https://orcid.org/0000-0002-1807-4644}{ORCID})

}
\keyword{internal}
