% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimorph_0.1_functions_dimorphism_uni_and_multi_2024-07-16.r
\name{dimorph}
\alias{dimorph}
\title{Estimate Dimorphism in a Univariate or Multivariate Sample}
\usage{
dimorph(
  x,
  method = "SSD",
  methodMulti = "GMM",
  sex = NULL,
  sex.female = 1,
  center = "geomean",
  ads = NULL,
  templatevar = NULL,
  na.rm = T,
  ncorrection = F,
  details = F,
  dfout = F
)
}
\arguments{
\item{x}{A dataframe, matrix, or vector of positive numbers corresponding to measurements for one or more size variable(s).  Values must be in the original measurement space, \bold{not} log-transformed.}

\item{method}{A character string specifying the univariate method used to calculate or estimate dimorphism.  Options 
  include:
  \itemize{
  \item \code{"SSD"}: \bold{Sexual Size Dimorphism}. Follows Smith (1999). Calculates actual sexual 
    dimorphism in the sample as the ratio of
    mean male size to mean female size.  Depending on \code{center}, sex-specific means are calculated
    either as geometric means or arithmetic means.  Requires \code{sex} to be specified.
  \item \code{"MMR"}: \bold{Mean Method Ratio}. Follows Godfrey et al. (1993). Splits the sample at its mean, 
    then calculates the ratio of the mean 
    of measurements larger than the overall mean to the mean of measurements smaller than the
    overall mean.  If any measurements are exactly equal to the overall mean, they contribute to both
    the larger and smaller group as half an individual in a weighted mean.  Depending on \code{center}, 
    the overall mean and subgroup means are calculated either as geometric means or arithmetic 
    means.  Ignores \code{sex}.
  \item \code{"BDI"}: \bold{Binomial Dimorphism Index}. Follows Reno et al. (2003). Given \emph{n} measurements,
    calculates all possible ratios of the mean of larger specimens to the mean of
    smaller specimens when the sample is split into the \emph{k} largest specimens and
    \emph{n-k} smallest specimens, where \emph{k} ranges from 1 to \emph{n}-1.  A weighted 
    mean is then calculated for all ratios, where the weights are equal to the probability
    of \emph{k} successes in \emph{n} trials in a binomial distribution. Depending on 
    \code{center}, means are calculated either as geometric means or arithmetic 
    means.  Ignores \code{sex}.
  \item \code{"ERM"}: \bold{Exact Resampling Method}.  Modification of Lee's (2001) Assigned Resampling 
    Method (ARM).  ARM is a resampling-based estimate of dimorphism that repeatedly samples two values
    with replacement from \code{x}, then calculates their ratio as long as both are neither more than 0.5 standard
    deviations above the mean or both 0.5 standard deviations below the mean (otherwise the pair is 
    rejected).  ARM typically oversamples the possible combination of two values sampled from a small sample 
    (as originally described it samples 1,000 pairs, whereas a sample of 20 measurements only has 210 possible pairs) 
    and sampling with replacement biases dimorphism estimates downwards by the incorporation of multiple ratios
    of 1 whever the same value is sampled twice and is not rejected by retention criterion.
    \code{"ERM"} performs an exact resampling of all possible
    pairs of values without replacement, but otherwise follows Lee's algorithm.
    Depending on \code{center}, the procedure is applied in either logarithmic (\code{"geomean"}) or
    raw (\code{"mean"}) data space.  Ignores \code{sex}.
  \item \code{"MoM"}: \bold{Method of Moments}.  Follows Josephson et al. (1996).  Assumes that the sample is 
    a mixture of two underlying lognormal distributions and uses three moments around the mean of the logged
    combined sex distribution to estimate the means of the underlying distributions.  This calculation is 
    always performed on the log-transformed data regardless of the value of \code{center}.
    Assumes that the sample contains an equal proportion of females and males and that those two subsamples 
    have equal variance.  Ignores \code{sex}.
  \item \code{"FMA"}: \bold{Finite Mixture Analysis}. Follows Godfrey et al. (1993). Assumes that the sample is 
    a mixture of two underlying normal or lognormal distributions.  Assumes that the sample contains an equal
    proportion of females and males and that those two subsamples have equal variance, then estimates the
    maximum separation of the two means. Depending on \code{center}, the underlying distributions are
    treated as either normal (\code{"mean"}) or lognormal (\code{"geomean"}).  Ignores \code{sex}.
  \item \code{"BFM"}: \bold{Bayesian Finite Mixture}.  Assumes that the sample is a finite mixture of two 
    underlying normal or lognormal distributions.  Unlike \code{"FMA"} and \code{"MoM"}, estimates 
    the proportion of females and males assuming that they may not be equal, and uses a Bayesian Information 
    Criterion (BIC) approach to select between a model that estimates a single variance for both sexes and a 
    model that estimates variances separately for the two constituent distributions using 
    \code{\link[mclust]{mclustBIC}}.  It then calculates the ratio of
    the two estimated means. Depending on \code{center}, the underlying distributions are
    treated as either normal (\code{"mean"}) or lognormal (\code{"geomean"}).  Ignores \code{sex}.
    When performed on lognormal data, this method is similar to the pdPeak method of Sasaki et al. (2021), 
    particularly when the BIC procedure selects an equal variance model (which it typically does).
  \item \code{"CV"}: \bold{Coefficient of Variation}.  Calculates the coefficient of variation as the 
    standard deviation of \code{x} divided by the mean of \code{x} then multiplied by 100.  This calculation
    is always performed on the raw data regardless of the value of \code{center} (an analogous 
    method using logarithmic data is \code{"sdlog"}).  Ignores \code{sex}.  Additionally, Sokal and Braumann's (1980) 
    size correction factor can be applied by setting \code{ncorrection} to \code{TRUE}, although this is
    \code{FALSE} by default.
  \item \code{"CVsex"}: \bold{Modified Coefficient of Variation}.  Calculates a modified 
    version of the coefficient of variation: the standard deviation is replaced by the 
    square root of the sum of squared differences of every value of \code{x} from the unweighted
    mean of the sex-specific means in \code{x} divided by the square root of \emph{n}-1, and this is 
    divided by the unweighted mean of the sex-specific means, then multiplied by 100.  This calculation
    is always performed on the raw data regardless of the value of \code{center} (an analogous 
    method using logarithmic data is \code{"sdlogsex"}).  Requires \code{sex}.  Additionally, Sokal and Braumann's (1980) 
    size correction factor can be applied by setting \code{ncorrection} to \code{TRUE}, although this is
    \code{FALSE} by default.
  \item \code{"sdlog"}: \bold{Standard Deviation of Logged Data}.  First, \code{x} is log-transformed
    using the natural logarithm, then the standard deviation is calculated.  This is a measure of
    proportional variation of the values of \code{x} about their geometric mean; \emph{i.e.}, analagous
    to the coefficient of variation for a lognormal distribution. This calculation
    is always performed on the log-transformed data regardless of the value of \code{center} 
    (an analogous method using raw data is \code{"CV"}).  Ignores \code{sex}.
  \item \code{"sdlogsex"} \bold{Modified Standard Deviation of Logged Data}.  First, \code{x} is 
    log-transformed using the natural logarithm.  Then a modified version of standard deviation is 
    calculated: the square root of the sum of squared differences of every logged value of \code{x} 
    from the unweighted mean of the sex-specific means of log-transformed \code{x}, divided by the 
    square root of \emph{n}-1. This calculation is always performed on the log-transformed data 
    regardless of the value of \code{center}  (an analogous method using raw data is \code{"CVsex"}).
    Requires \code{sex}.
} Defaults to \code{"SSD"}.}

\item{methodMulti}{A character string specifying the multivariate method used to 
  calculate or estimate dimorphism.  Note that regardless of the value of this argument,
  multivariate estimation procedures will only be carried out if \code{x} is a multivariate 
  dataset.  Options include:
  \itemize{
  \item \code{"GMM"}: Follows the Geometric Mean Method of Gordon et al. (2008).  The
    selected univariate method is applied to all variables, then the geometric mean
    is calculated of the dimorphism estimates for all variables to produce a single
    estimate for the whole data set. Note 
    that this methodology is not appropriate for variance-based univariate methods;
    \emph{i.e.}, \code{"CV"}, \code{"CVsex"}, \code{"sdlog"}, and \code{"sdlogsex"}.
  \item \code{"GMsize"}:  If \code{x} is a dataframe or matrix, this method first 
    calculates overall size as the geometric mean of measurements in all variables for
    those specimens that are complete for all variables in the data set.  The selected
    univariate method is then applied to this measure of overall size.
  \item \code{"TM"}:  Follows the template method of Reno et al. (2003). A variable of interest
    is specified by the user with the argument \code{templatevar}.  The algorithm identifies
    a template individual that can be used to estimate the largest number of values for 
    the selected variable of interest by using ratios between the value of that variable and
    other variables in the template individual, which are then multiplied by the value of those
    other variables in individuals missing the target variable to maximixe the dataset for 
    that variable.  A user-selected univariate method is then applied to the combined dataset
    of actual and estimated values for the target variable.
} Defaults to \code{"GMM"}.}

\item{sex}{A vector indicating sex for the measurements in \code{x}.  If present, must include 
exactly two groups and have the same length as the number of specimens in \code{x}.  Non-factor 
vectors will be coerced to factors if possible.  May be \code{NULL} since some methods do not 
require sex information.  Methods which require sex information will generate an error if 
\code{sex} is \code{NULL}.  For methods that do not require sex information, if sex is provided 
it will be ignored for the calculation of the estimate, but it will be used to report the actual 
proportion of females and males in the sample.  Defaults to \code{NULL}.}

\item{sex.female}{An integer scalar (1 or 2) specifying which level of \code{sex} 
corresponds to female.  Ignored if \code{sex} is \code{NULL}.  Defaults to 1.}

\item{center}{A character string specifying the method used to calculate a mean, either \code{"geomean"} 
(default) which uses the geometric mean, or \code{"mean"} which uses the arithmetic mean.  More broadly, 
\code{"geomean"} indicates analyses are conducted in logarithmic data space and \code{"mean"} indicates 
analyses are conducted in raw data space.  Some methods can only be applied in one domain or the other: 
\code{"CV"} and \code{"CVsex"} are always calculated in raw data space and \code{center} will be set to 
\code{"mean"} for these methods regardless of the value set by the user; \code{"MoM"}, \code{"sdlog"}, 
and \code{"sdlogsex"} are always calculated in logarithmic data space and \code{center} will be set to 
\code{"geomean"} for these methods regardless of the value set by the user.}

\item{ads}{A vector of integer addresses for positions in the data vector \code{x} to be included 
in the calculation of dimorphism; any other data in \code{x} will be ignored.  If \code{ads} is
\code{NULL} then all data are included in the calculation.  Defaults to \code{NULL}.}

\item{templatevar}{A character object or integer value specifying the name or column number 
of the variable in \code{x} to be estimated using the template method.  Ignored if template method 
is not used.  Defaults to \code{NULL}.}

\item{na.rm}{A logical scalar indicating whether NA values should be stripped before
the computation proceeds.  Defaults to \code{TRUE}.}

\item{ncorrection}{A logical scalar indicating whether to apply Sokal and Braumann's (1980) 
size correction factor to CV estimates.  Defaults to \code{FALSE}.}

\item{details}{A logical scalar indicating whether variable name and specimen names 
should be retained (if available) as attributes in the output object.  Defaults to 
\code{FALSE}.}

\item{dfout}{A logical scalar indicating whether the result should be given as a
\code{dimorphEstDF} object; if \code{FALSE}, returns a \code{dimorphEst} object.  Defaults to \code{FALSE}.}
}
\value{
Either a class \code{dimorphEst} or \code{dimorphEstDF} object. \code{dimorphEst} objects are numeric 
  vectors of length one corresponding to measured or estimated dimorphism in \code{x} with associated information 
  preserved as attributes.  \code{dimorphEstDF} objects are single-row data frames that contain the dimorphism 
  estimate for \code{x} along with other associated information.  Applying \code{summary} to either of these 
  objects provides information about the dataset and method used to generate it.
}
\description{
Function to calculate or estimate dimorphism for a univariate or multivariate sample.
}
\examples{
## Univariate estimates:
data(apelimbart)
gorillas <- apelimbart[apelimbart$Species=="Gorilla gorilla",]
# Next line would generate an error: sex is required
# dimorph(x=gorillas$FHSI, method="SSD") 
gorSSD  <- dimorph(x=gorillas$FHSI, # variable and specimen names not preserved
                   method="SSD", sex=gorillas$Sex, details=TRUE)
gorSSD2 <- dimorph(x=gorillas[,"FHSI",drop=FALSE], # variable and specimen names preserved
                   method="SSD", sex=gorillas$Sex, details=TRUE)
gorSSD
str(gorSSD)
summary(gorSSD)
summary(gorSSD2) # results are identical to 'gorSSD'
summary(gorSSD, verbose=TRUE)
summary(gorSSD2, verbose=TRUE) # variable and specimen names preserved
# A subset of specimens can be specified for analysis using 'ads'
summary(dimorph(x=gorillas$FHSI, method="SSD", sex=gorillas$Sex, ads=c(1:10, 51:60)))
# Methods for estimating dimorphism:
summary(dimorph(x=gorillas$FHSI, method="MMR"))
summary(dimorph(x=gorillas$FHSI, method="BDI"))
summary(dimorph(x=gorillas$FHSI, method="MoM"))
summary(dimorph(x=gorillas$FHSI, method="FMA"))
summary(dimorph(x=gorillas$FHSI, method="BFM"))
summary(dimorph(x=gorillas$FHSI, method="ERM"))
summary(dimorph(x=gorillas$FHSI, method="CV"))
summary(dimorph(x=gorillas$FHSI, method="CV", ncorrection=TRUE))
summary(dimorph(x=gorillas$FHSI, method="CVsex", sex=gorillas$Sex))
summary(dimorph(x=gorillas$FHSI, method="CVsex", sex=gorillas$Sex, ncorrection=TRUE))
summary(dimorph(x=gorillas$FHSI, method="sdlog"))
summary(dimorph(x=gorillas$FHSI, method="sdlogsex", sex=gorillas$Sex))

# Now setting 'dfout' to TRUE:
allmethods <- rbind(dimorph(x=gorillas$FHSI, method="SSD", sex=gorillas$Sex, dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="MMR", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="BDI", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="MoM", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="FMA", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="BFM", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="ERM", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="CV", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="CVsex", sex=gorillas$Sex, dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="sdlog", dfout=TRUE),
                    dimorph(x=gorillas$FHSI, method="sdlogsex", sex=gorillas$Sex, dfout=TRUE))				
allmethods

# Alternatively, using apply
res <- apply(data.frame(method=c("SSD","MMR","BDI","MoM","FMA","BFM",
                                 "ERM","CV","CVsex","sdlog","sdlogsex")),
             MARGIN=1,
             FUN=function(method, ...) dimorph(x=gorillas$FHSI, method=method,
                                               sex=gorillas$Sex, dfout=TRUE),
             simplify=FALSE)
as.data.frame(do.call(rbind, res))

## Multivariate estimates:
# GMsize (only usable for complete datasets)
Gg.GMsize <- dimorph(x=gorillas[,c("FHSI","HHMaj","TPMAP","RHMaj")],
                     method="SSD", methodMulti="GMsize", sex=gorillas$Sex, details=TRUE)
Gg.GMsize
summary(Gg.GMsize)

# GMM (produces the same values for ratio estimators as GMsize when data are complete)
Gg.GMM1 <- dimorph(x=gorillas[,c("FHSI","HHMaj","TPMAP","RHMaj")],
                   method="SSD", methodMulti="GMM", sex=gorillas$Sex, details=TRUE)
# now with subset of gorilla data
Gg.GMM2 <- dimorph(x=gorillas[,c("FHSI","HHMaj","TPMAP","RHMaj")],
                   method="SSD", methodMulti="GMM", sex=gorillas$Sex,
                   ads=c(1:10, 51:60), details=TRUE)
summary(Gg.GMM1)
summary(Gg.GMM2)

## Now with some simulated fossil data
SSDvars <- c("FHSI", "TPML", "TPMAP", "TPLAP", "HHMaj",
             "HHMin", "RHMaj", "RHMin", "RDAP", "RDML")
Fs1 <- fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]
Fs1GMM <- dimorph(x=Fs1, method="MMR", methodMulti="GMM", details=TRUE)
Fs1TM <- dimorph(x=Fs1, method="MMR", methodMulti="TM", 
                 templatevar="FHSI", details=TRUE)
Fs1GMM
Fs1TM
summary(Fs1GMM)
summary(Fs1TM, verbose=TRUE)

Fs1Both <- rbind(dimorph(x=Fs1, method="MMR", methodMulti="GMM", details=TRUE, dfout=TRUE),
                 dimorph(x=Fs1, method="MMR", methodMulti="TM", 
                         templatevar="FHSI", details=TRUE, dfout=TRUE))
Fs1Both
}
\references{
Godfrey LR, Lyon SK, Sutherland MR. (1993) Sexual dimorphism in large-bodied 
  primates: The case of the subfossil lemurs. \emph{American Journal of Physical 
  Anthropology}. 90:315-334. (\href{https://doi.org/10.1002/ajpa.1330900306}{https://doi.org/10.1002/ajpa.1330900306})

Gordon AD, Green DJ, Richmond BG. (2008) Strong postcranial size dimorphism in 
  \emph{Australopithecus afarensis}: Results from two new resampling methods for multivariate 
  data sets with missing data. \emph{American Journal of Physical Anthropology}. 135:311-328.
  (\href{https://doi.org/10.1002/ajpa.20745}{https://doi.org/10.1002/ajpa.20745})

Josephson SC, Juell KE, Rogers AR. (1996) Estimating sexual dimorphism by 
  method-of-moments. \emph{American Journal of Physical Anthropology}. 100:191-206. 
  (\href{https://doi.org/10.1002/(SICI)1096-8644(199606)100:2<191::AID-AJPA3>3.0.CO;2-0}{https://doi.org/10.1002/(SICI)1096-8644(199606)100:2<191::AID-AJPA3>3.0.CO;2-0})

Lee S-H. (2001) Assigned Resampling Method: A new method to estimate size sexual dimorphism in 
  samples of unknown sex. \emph{Anthropological Review}. 64:21–39. 
  (\href{https://doi.org/10.18778/1898-6773.64.02}{https://doi.org/10.18778/1898-6773.64.02})

Reno PL, Meindl RS, McCollum MA, Lovejoy CO. (2003) Sexual dimorphism in 
  \emph{Australopithecus afarensis} was similar to that of modern humans. \emph{Proceedings 
  of the National Academy of Sciences}. 100:9404-9409. 
  (\href{https://doi.org/10.1073/pnas.1133180100}{https://doi.org/10.1073/pnas.1133180100})

Sasaki T, Semaw S, Rogers MJ, Simpson SW, Beyene Y, Asfaw B, White TD, Suwa G. (2021) Estimating sexual 
  size dimorphism in fossil species from posterior probability densities. \emph{Proceedings 
  of the National Academy of Sciences}. 118:e2113943118.
  (\href{https://doi.org/10.1073/pnas.2113943118}{https://doi.org/10.1073/pnas.2113943118})

Sokal RR, Braumann CA (1980) Significance tests for coefficients of variation and variability profiles. 
  \emph{Systematic Zoology}. 29:50–66. 
  (\href{https://doi.org/10.2307/2412626}{https://doi.org/10.2307/2412626})

Smith RJ. (1999) Statistics of sexual size dimorphism. \emph{Journal of 
  Human Evolution}. 36:423-458. (\href{https://doi.org/10.1006/jhev.1998.0281}{https://doi.org/10.1006/jhev.1998.0281})
}
\seealso{
\code{\link[mclust]{mclustBIC}},  \code{\link[dimorph]{SSDtest}}
}
