% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_0.1_2024-12-28.r
\docType{data}
\name{FMAtable}
\alias{FMAtable}
\title{Finite Mixture Analysis table from Godfrey et al. (1993)}
\format{
An object of class \code{data.frame} with 99 rows and 2 columns.
}
\usage{
data(FMAtable)
}
\description{
Values of \emph{k} for a given sample size for the calculation of distribution
  means for a Gaussian finite mixture using equations in Godfrey et al. (1993). \emph{k}
  gives the mean expected number of standard deviations (as measured for the whole dataset)
  represented in the observed data range for a given sample size.
}
\examples{
data(FMAtable)
FMAtable
}
\references{
Godfrey LR, Lyon SK, Sutherland MR. (1993) Sexual dimorphism in large-bodied 
  primates: The case of the subfossil lemurs. \emph{American Journal of Physical 
  Anthropology}. 90:315-334. (\href{https://doi.org/10.1002/ajpa.1330900306}{https://doi.org/10.1002/ajpa.1330900306})
}
\keyword{datasets}
