% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimorph 0.1 functions analyses 2024-07-01.r
\name{getstructure}
\alias{getstructure}
\title{Get the Structure Component for Dimorphism Resampling}
\usage{
getstructure(x, forcematrix = F)
}
\description{
Given a list of metric datasets, calculates the shared data structure that allows
for all samples to be included in a common resampling analysis.  This generates
the \code{struc} component to be included in various functions in the \code{dimorph}
package.
}
\examples{
SSDvars <- c("FHSI", "TPML", "TPMAP", "TPLAP", "HHMaj",
             "HHMin", "RHMaj", "RHMin", "RDAP", "RDML")
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars]))
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]))
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]), forcematrix=T)
}
\seealso{
\code{\link{SSDtest}}
}
