% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimorph_0.1_functions_plot_print_summary_2025-11-02.r
\name{plot.SSDtest}
\alias{plot.SSDtest}
\title{Plot \code{SSDtest} Object}
\usage{
\method{plot}{SSDtest}(
  x,
  est = 1,
  type = "est",
  diffs = NULL,
  nbins = 100,
  plottitle = NULL,
  groupcols = NULL,
  leg = T,
  legpos = NULL,
  legsize = 1,
  legtitle = "",
  lty = "dashed",
  titlesize = 2,
  invert = NULL,
  xlim = NULL,
  ylim = NULL,
  diffcol = "#CCB03D",
  gridplot = TRUE,
  useradvance = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{SSDtest} object.}

\item{est}{An integer specifiying the specific combination of methods (e.g., univariate 
method, centering algorithm) for the resampled estimates to be plotted.  This integer corresponds to 
the row number in the table providing mean or median estimates by method combination when printing
the \code{SSDtest} object.}

\item{type}{A character string specifying the type of plot to produce. \code{"est"} generates a plot
of histograms of estimates for all of the samples in \code{x}. \code{"diff"} gereates a plot of one 
or more histograms of all possible pairwise differences in estimates between pairs of samples, with 
the argument \code{diffs} specifying which pairs to plot.  Defaults to \code{"est"}.  Also, because 
the plotting function calculates histograms
for \code{type="diff"} by first calculating the difference between all possible pairs of one estimate taken 
from the two distributions of resampled values to be compared, plots using \code{type="diff"} can take a 
long time to be generated.  The output of \code{plot.SSDtest()} can be saved and plotted later without the need 
for recalculation.}

\item{diffs}{Specifies which set of samples to plot pairwise differences for if \code{type} is \code{"diff"}. If set
to \code{"all"} it generates a single plot with all possible pairs.  If given a vector of two or more integers,
it plots all of the differences for that set of samples (e.g., setting \code{diffs} to \code{c(1,3,4)} will plot 
three histograms: one for the differences in estimates between the first and third sample in \code{x}, one for the
differences between the first and fourth sample, and one for the differences between the third and fourth sample).
Any of the samples identified in \code{diffs} that do not include data for the estimate specified in \code{est} 
will not be included in pairwise comparisons (e.g., if \code{"SSD"} was included as a univariate method on a set 
of samples for which some had sex data and others didn't, the samples without sex data would be missing estimates 
for that method and thus would be excluded from pairwise comparisons of \code{"SSD"} estimates.) An error will
be generated in cases where fewer then two samples with estimates are specified.  Defaults to "all".}

\item{nbins}{An integer specifiying the number of bins to sort data into for histograms.}

\item{plottitle}{A character string specifying a title to print at the top of the plot.  If \code{NULL}, no 
title is printed.  Defaults to \code{NULL}.}

\item{groupcols}{A vector of hexidecimal colors of length equal to the number of samples in \code{x} for use
in plots when \code{type} is \code{"est"}.  Defaults to a palette generated by \code{\link[pals]{parula}}.}

\item{leg}{A logical scalar indicating whether a legend should be provided.  Defaults to \code{TRUE}.}

\item{legpos}{A character string specifying where to plot the legend.  Acceptable values are \code{"left"}, 
\code{"top"}, \code{"right"}, and \code{"bottom"}.  Defaults to \code{"right"}.}

\item{legsize}{A numeric scalar specifying a multiplier for the legend font size relative to the default font size. 
Defaults to \code{1}.}

\item{legtitle}{A character string specifying a title to print at the top of the legend.  If \code{NULL}, no 
title is printed.  Defaults to \code{NULL}.}

\item{lty}{Line type for plotting observed fossil value in scenario where a single point estimate is calculated
for a single fossil sample.  Defaults to \code{"dashed"}.}

\item{titlesize}{A numeric scalar specifying a multiplier for the title font size relative to the default font size. 
Defaults to \code{2}.}

\item{invert}{An integer vector specifying the group number of histograms to invert and project below the zero 
count line, where group number is determined by the order of datasets provided first in \code{fossil}, then in 
\code{comp}.  If \code{NULL}, no histograms are inverted.  Defaults to \code{NULL}.}

\item{xlim}{A numeric vector with two elements specifying the minimum and maximum values of the x-axis.  If 
\code{xlim} is \code{NULL}, default values generated by \code{\link[ggplot2]{ggplot}} are used.}

\item{ylim}{A numeric vector with two elements specifying the minimum and maximum values of the y-axis.  If 
\code{ylim} is \code{NULL}, default values generated by \code{\link[ggplot2]{ggplot}} are used.}

\item{diffcol}{A color used to fill in histogram bars for differences in parameter values between two samples. 
defaults to \code{"#CCB03D"}.}

\item{gridplot}{A logical scalar indicating whether the multiple plots generated by \code{type="diff"} should 
be plotted as a single grid of plots or not.  If \code{FALSE} then individual plots are printed sequentially. 
Defaults to \code{TRUE}.}

\item{useradvance}{A logical scalar indicating whether to prompt the user between plots when multiple plots 
are generated by \code{type="diff"} and \code{gridplot} is \code{FALSE}. If \code{FALSE} then the function 
waits two seconds between plots. Defaults to \code{TRUE}.}

\item{...}{Arguments to be passed to other functions.  Not currently used.}
}
\description{
Plot the object output by \code{\link[dimorph]{SSDtest}}.  \strong{N.B.:} warning messages stating 
  "\code{Removed} \emph{n} \code{rows containing  missing values (`geom_bar()`).}" are generated 
  for any histogram that contains bars of height zero, which will be true for most histograms.  This is 
  due to a known issue with how \code{\link[ggplot2]{ggplot}} handles histograms (see 
  \href{https://github.com/tidyverse/ggplot2/issues/3265}{https://github.com/tidyverse/ggplot2/issues/3265}), 
  and these warning messages can safely be ignored.
}
\examples{
SSDvars <- c("FHSI", "TPML", "TPMAP", "TPLAP", "HHMaj", "HHMin", "RHMaj", "RHMin", "RDAP", "RDML")
test_faux_multi1 <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", SSDvars],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", SSDvars]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     methsMulti=c("GMM"),
     datastruc="both",
     nResamp=100,
     templatevar="FHSI")
test_faux_multi1
plot(test_faux_multi1)
speciescolors <- c("Fauxil sp. 1"="#352A87", "G. gorilla"="#EABA4B",
                   "H. sapiens"="#09A9C0", "P. troglodytes"="#78BE7C",
                   "H. lar"="#0D77DA")
# change colors of sample distributions                   
plot(test_faux_multi1, groupcols=speciescolors)
# plot estimates for the fourth methcombo (GMM and MMR)
plot(test_faux_multi1, est=4, groupcols=speciescolors)
# invert second and fourth sample distributions
plot(test_faux_multi1, est=4, groupcols=speciescolors, invert=c(2,4))
# plot distributions of differences among sample estimates for method combination 4
plot(test_faux_multi1, est=4, type="diff")
# plot differences between second and third samples
plot(test_faux_multi1, est=4, type="diff", diffs=c(2,3)) 
}
