% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimorph_0.1_functions_general_2024-11-02.r
\name{pvals}
\alias{pvals}
\title{p-Values for \code{SSDtest} Object}
\usage{
pvals(x, alternative = "two.sided")
}
\arguments{
\item{x}{An object of class \code{SSDtest}.}

\item{alternative}{a character value indicating whether two-sided or one-sided p-values should be extracted. 
Takes \code{"two.sided"} or \code{"one.sided"}, defaulting to \code{"two.sided"}.  If \code{"one.sided"} 
then both sets of one-sided p-values are returned.}
}
\value{
A matrix containing the corresponding p-values in \code{x}.
}
\description{
Function for extracting the two-sided or one-sided p-values in a \code{SSDtest} object as a matrix.
}
\examples{
# Standard significance test with one fossil sample, sampling without replacement
test_faux_uni <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI"]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI"],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI"],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "FHSI"],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI"]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("MMR", "BDI"),
     limit=1000,
     nResamp=100)
pvals(test_faux_uni) # defaults to two-sided
pvals(test_faux_uni, alternative="one.sided")
}
