% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimorph_0.1_functions_bootstrap_and_specific_resampling_procedures_2024-07-16.r
\name{SSDtest}
\alias{SSDtest}
\title{Significance Tests for Sexual Size Dimorphism Estimates}
\usage{
SSDtest(
  fossil = NULL,
  comp = NULL,
  fossilsex = NULL,
  compsex = NULL,
  methsUni = c("MMR", "BDI"),
  methsMulti = "GMM",
  replace = F,
  rebootstrap = F,
  fullsamplesboot = F,
  nResamp = 1000,
  exactcomp = T,
  exactfossil = T,
  limit = 50000,
  center = "geomean",
  sex.female = 1,
  na.rm = T,
  ncorrection = F,
  matchvars = F,
  datastruc = NULL,
  templatevar = NULL
)
}
\arguments{
\item{fossil}{A list of matrices or data frames of measurements from fossil sample(s) (or other test samples) to be
compared in a series of pairwise resampling tests, with rows corresponding to individual specimens and 
columns corresponding to size variables.  Sex data should not be included.  Some data can be missing. 
When a single fossil sample is present, by default null distributions of resampled values from comparative data sets 
will be compared to a single point estimate generated for the fossil sample (one point estimate per estimation 
method).  \code{fossil} may be \code{NULL}, in which case all samples should be provided in \code{comp}, and all 
samples will be resampled.}

\item{comp}{A list of matrices or data frames of measurements from comparative sample(s) to be
compared in a series of pairwise resampling tests, with rows corresponding to individual specimens and 
columns corresponding to size variables.  Sex data should not be included.  All data sets must be complete
for all measurements.}

\item{fossilsex}{A list of vectors indicating sex for the individuals in each of the samples in \code{fossil}. 
Defaults to \code{NULL}.}

\item{compsex}{A list of vectors indicating sex for the individuals in each of the samples in \code{comp}. 
Defaults to \code{NULL}.}

\item{methsUni}{A character vector specifying the univariate method(s) used to calculate or estimate dimorphism. 
See \code{\link{dimorph}} for options.  Defaults to \code{c("MMR", "BDI")}.}

\item{methsMulti}{A character vector specifying the multivariate method(s) used to 
calculate or estimate dimorphism.  Note that regardless of the value of this argument,
multivariate estimation procedures will only be carried out if \code{fossil} and \code{comp} are multivariate 
datasets.  See \code{\link{dimorph}} for options.  Defaults to \code{"GMM"}.}

\item{replace}{Logical scalar specifying whether to sample from comparative datasets with replacement or not.
Defaults to \code{FALSE}.}

\item{rebootstrap}{Logical scalar specifying whether to add an additional step after initial resampling in 
which resampled addresses for both comparative and fossil datasets are bootstrapped (i.e., sampled with 
replacement to an equal sample sizes as the initial set of resampled addresses).  This procedure implements 
the highly conservative "resampled extinct distribution method" of Gordon et al. (2008). Defaults to 
\code{FALSE}.  Warning: setting \code{rebootstrap} to \code{TRUE} will drastically reduce power!}

\item{fullsamplesboot}{Logical scalar.  If all samples are complete (no missing data) and \code{fullsamplesboot} is 
set to \code{TRUE}, rather than downsampling to the minimum sample in \code{fossil}, all samples in \code{fossil} are
and \code{comp} are bootstrapped (i.e., sampled with replacement at their full sample size).  Defaults to 
\code{FALSE}.  Note that for univariate analyses, \code{NA}s are removed by default so all datasets (including 
fossils) are considered complete.  Therefore setting \code{fullsamplesboot} to \code{TRUE} for univariate analyses 
with fossils will bootstrap the fossil dataset rather than generating a single point estimate from the fossil dataset.}

\item{nResamp}{Integer specifying the number of resampling iterations 
if Monte Carlo sampling is used.  Defaults to \code{1,000}.}

\item{exactcomp}{Logical scalar specifying for samples in \code{comp} whether to sample all possible unique 
combinations of resampled datasets for the minimum data structure present in \code{fossil}. 
If set to \code{FALSE}, or if set to \code{TRUE} and the number of unique combinations exceeds \code{limit}, 
then Monte Carlo sampling is used instead.  Defaults to \code{TRUE}.}

\item{exactfossil}{Logical scalar specifying for samples in \code{fossil} whether to sample all 
possible unique combinations of resampled datasets for the minimum data structure present in \code{fossil}. 
If set to \code{FALSE}, or if set to \code{TRUE} and the number of unique combinations exceeds \code{limit}, 
then Monte Carlo sampling is used instead.  Defaults to \code{TRUE}.}

\item{limit}{Integer setting the upper limit on the number of unique combinations allowable 
for exact resampling.  If exact resampling would produce more resampled datasets than this number, 
Monte Carlo resampling is used instead.  Defaults to \code{50,000}.}

\item{center}{A character string specifying the method used to calculate a mean, either \code{"geomean"} 
(default) which uses the geometric mean, or \code{"mean"} which uses the arithmetic mean.  More broadly, 
\code{"geomean"} indicates analyses are conducted in logarithmic data space and \code{"mean"} indicates 
analyses are conducted in raw data space.  Some methods can only be applied in one domain or the other: 
\code{"CV"} and \code{"CVsex"} are always calculated in raw data space and \code{center} will be set to 
\code{"mean"} for these methods regardless of the value set by the user; \code{"MoM"}, \code{"sdlog"}, 
and \code{"sdlogsex"} are always calculated in logarithmic data space and \code{center} will be set to 
\code{"geomean"} for these methods regardless of the value set by the user.}

\item{sex.female}{An integer scalar (1 or 2) specifying which level of \code{sex} 
corresponds to female.  Ignored if \code{sex} is \code{NULL}.  Defaults to 1.}

\item{na.rm}{A logical scalar indicating whether \code{NA} values should be stripped before
the computation proceeds in univariate analyses.  Not relevant for multivariate analyses.  Defaults to \code{TRUE}.}

\item{ncorrection}{A logical scalar indicating whether to apply Sokal and Braumann's (1980) 
size correction factor to CV estimates.  Defaults to \code{FALSE}.}

\item{matchvars}{Logical scalar specifying whether to compare the shared set of variable names in
\code{comp} and \code{fossil} to the variable names in \code{struc} and pare them all down to the 
set of shared variables. If \code{FALSE} and variable names differ then an error will be returned.
Defaults to \code{FALSE}.}

\item{datastruc}{If multivariate data are used, this is a character string specifiying whether to 
incorporate the missing data structure
into dimorphism estimates (\code{"missing"}), whether to downsample to the missing data sample size but 
keep all metric data for the comparative sample (\code{"complete"}), or to perform both types of
resampling separately (\code{"both"}).  Ignored if only univariate data are provided or if all datasets
are complete. Defaults to \code{NULL}, which reverts to \code{"missing"} if some datasets are incomplete.}

\item{templatevar}{A character object or integer value specifying the name or column number 
of the variable in \code{fossil} and \code{comp} to be estimated using the template method.  Ignored 
if template method is not used.  Defaults to \code{NULL}.}
}
\value{
A list of class \code{SSDtest}.  Printing this object provides information about the datasets and 
  test performed, median values for resampled distributions for each sample for each method used, confidence 
  intervals for all estimators calculated, and calculated \emph{p}-values.  Plotting this object produces 
  histograms for resampled distributions for one or more estimation methods, or histograms for differences
  between samples (see \code{\link[dimorph]{plot.SSDtest}}).
}
\description{
Function to calculate \emph{p}-values for pairwise comparisons of dimorphism estimates for two or more univariate 
  or multivariate samples.  The first step in this process identifies the minimum sample size (and missing data 
  structure) present in the samples of \code{fossil}; for multivariate datasets with missing data, datasets are 
  restricted to individuals with measurements in variables present in all samples.  All datasets are then resampled to
  the sample size (and missing data structure) of the minimum sample.  This function generates both
  two-sided and one-sided \emph{p}-values for each pair of samples.
}
\examples{
# SSDtests using simulated fossils generated from real gorilla and human samples with some 
#   data removed
data(fauxil)
fauxil

## Univariate examples
# First, some code that would generate errors
# SSDtest()
# SSDtest(fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI"]))
# SSDtest(comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI"]))

# Standard significance test with one fossil sample, sampling without replacement
test_faux_uni <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI"]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI"],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI"],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "FHSI"],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI"]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     limit=1000,
     nResamp=100)
test_faux_uni
plot(test_faux_uni) # plots first method by default (SSD) - warning due to known ggplot issue
plot(test_faux_uni, est=2) # plots second method (MMR)
speciescolors <- c("Fauxil sp. 1"="#352A87", "Fauxil sp. 2"="#F9FB0E", "G. gorilla"="#EABA4B",
                   "H. sapiens"="#09A9C0", "P. troglodytes"="#78BE7C", "H. lar"="#0D77DA")
plot(test_faux_uni, est=2, groupcols=speciescolors) # change the colors
plot(test_faux_uni, type="diff", est=2) # plots differences between samples
plot(test_faux_uni, type="diff", est=2, diffs=c(1,2)) # plots diffs between first pair of samples

# Same as above, but variable name information is preserved
test_faux_uni2 <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI", drop=FALSE]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI", drop=FALSE],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI", drop=FALSE],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes","FHSI",drop=FALSE],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI", drop=FALSE]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     limit=1000,
     nResamp=100)
test_faux_uni2
plot(test_faux_uni2, est=2, groupcols=speciescolors)

# Same as above except that null distributions are generated WITH replacement rather than WITHOUT
test_faux_uni3 <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI", drop=FALSE]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI", drop=FALSE],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI", drop=FALSE],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes","FHSI",drop=FALSE],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI", drop=FALSE]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     limit=1000,
     nResamp=100,
     replace=TRUE)
test_faux_uni3
plot(test_faux_uni3, est=2, groupcols=speciescolors)

# Instead of standard test, each comparative and fossil sample is bootstrapped (sampled 
#   with replacement to that sample's full sample size) by setting 'fullsamplesboot' to
#   TRUE.  Produces a distribution of fossil estimates rather than a single point estimate, 
#   but note that comparative samples are resampled to their full sample size, not 
#   downsampled to the fossil sample size.
test_faux_uni4 <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", "FHSI", drop=FALSE]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "FHSI", drop=FALSE],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "FHSI", drop=FALSE],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes","FHSI",drop=FALSE],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "FHSI", drop=FALSE]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     limit=1000,
     nResamp=100,
     fullsamplesboot=TRUE)
test_faux_uni4
plot(test_faux_uni4, est=2, # plots the second estimation method (MMR)
     invert=1, # inverts the histogram for the first group in 'test_afar_uni4'
     groupcols=speciescolors)

## Multivariate examples
# GMM significance tests with a fossil sample for multiple estimators using both complete
#   and incomplete comparative datasets.  A single point estimate is generated for the fossil 
#   sample for each method.
SSDvars <- c("FHSI", "TPML", "TPMAP", "TPLAP", "HHMaj",
             "HHMin", "RHMaj", "RHMin", "RDAP", "RDML")
test_faux_multi1 <- SSDtest(
     fossil=list("Fauxil sp. 1"=fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", SSDvars],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", SSDvars]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     methsMulti=c("GMM"),
     datastruc="both",
     nResamp=100,
     templatevar="FHSI")
test_faux_multi1
plot(test_faux_multi1, groupcols=speciescolors)
plot(test_faux_multi1, est=4, # plot the 4th method combination: MMR, GMM, missing datastructure
     groupcols=speciescolors)
plot(test_faux_multi1, est=4,
     type="diff") # plot the diffs between samples for estimates using the 4th method combination

# As above but with a different fossil sample.
test_faux_multi2 <- SSDtest(
     fossil=list("Fauxil sp. 2"=fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", SSDvars],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", SSDvars]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     methsMulti=c("GMM"),
     datastruc="both",
     nResamp=100,
     templatevar="FHSI")
test_faux_multi2
plot(test_faux_multi2, groupcols=speciescolors)
plot(test_faux_multi2, est=4, groupcols=speciescolors)
plot(test_faux_multi2, est=4, type="diff")

# Now the same data using a much more conservative approach by setting 'rebootstrap' to TRUE.
test_faux_multi3 <- SSDtest(
     fossil=list("Fauxil sp. 2"=fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]),
     comp=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
               "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
               "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", SSDvars],
               "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", SSDvars]),
     fossilsex=NULL,
     compsex=list("G. gorilla"=apelimbart[apelimbart$Species=="Gorilla gorilla", "Sex"],
                  "H. sapiens"=apelimbart[apelimbart$Species=="Homo sapiens", "Sex"],
                  "P. troglodytes"=apelimbart[apelimbart$Species=="Pan troglodytes", "Sex"],
                  "H. lar"=apelimbart[apelimbart$Species=="Hylobates lar", "Sex"]),
     methsUni=c("SSD", "MMR", "BDI"),
     methsMulti=c("GMM"),
     datastruc="both",
     nResamp=100,
     rebootstrap=TRUE,
     templatevar="FHSI")
test_faux_multi3
plot(test_faux_multi3, est=3, invert=1, groupcols=speciescolors)
}
\references{
Gordon AD, Green DJ, Richmond BG. (2008) Strong postcranial size dimorphism in 
  \emph{Australopithecus afarensis}: Results from two new resampling methods for multivariate 
  data sets with missing data. \emph{American Journal of Physical Anthropology}. 135:311-328.
  (\href{https://doi.org/10.1002/ajpa.20745}{https://doi.org/10.1002/ajpa.20745})
}
\seealso{
\code{\link[dimorph]{plot.SSDtest}}
}
