% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimorph_0.1_functions_analyses_2025-01-08.r
\name{getstructure}
\alias{getstructure}
\title{Get the Structure Component for Dimorphism Resampling}
\usage{
getstructure(x, forcematrix = F)
}
\arguments{
\item{x}{A list of one or more dataframes corresponding to samples of metric data where
rows are individual specimens and columns are variables.  Missing data are allowable
and should be indicated with \code{NA}s.}

\item{forcematrix}{A logical flag specifying whether a dataframe (instead of a vector of
specimen counts) should be returned when \code{x} only contains a single dataset and 
that dataset has missing data.  See "Value" below for more details.}
}
\value{
If \code{x} is a list of one or more datasets with no missing data, a dataframe
  filled with \code{1}s with a number of rows equal to the minimum number of specimens
  across all datasets in \code{x}, and columns corresponding to each variable shared
  by all datasets in \code{x}. If \code{x} contains one or more datasets with no missing data 
  and only one dataset with missing data, \code{getstructure} returns a dataframe equal to
  the dataset with missing data where all measurements are replaced with \code{1}, and only 
  variables shared across all datasets in \code{x} are present.  If \code{x} contains one 
  or more datasets with no missing data and two or more datasets with missing data, \code{getstructure} 
  returns a vector of integers in which each entry corresponds to a variable shared by
  all datasets in \code{x}, and the value of each entry is the minimum number of specimens with
  data for that variable across all datasets in \code{x}.  If \code{x} contains only a single dataset
  and that dataset has missing data, \code{getstructure} returns an integer vector as described above
  unless \code{forcematrix} is set to \code{TRUE}, in which case a dataframe is returned which is
  equal to the dataset in \code{x} where all measurements are replaced with \code{1}.  Note that
  an error will be generated if one or more complete datasets are present in \code{x} and any of 
  the complete datasets have fewer individuals than any datasets with missing data because the 
  complete data cannot be sampled down to the size of the dataset(s) with missing data.
}
\description{
Given a list of metric datasets, calculates the shared data structure that allows
for all samples to be included in a common resampling analysis.  This generates
the \code{struc} component to be included in various functions in the \code{dimorph}
package.
}
\examples{
SSDvars <- c("FHSI", "TPML", "TPMAP", "TPLAP", "HHMaj",
             "HHMin", "RHMaj", "RHMin", "RDAP", "RDML")
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars]))
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]))
getstructure(list(apelimbart[apelimbart$Species=="Gorilla gorilla", SSDvars],
                  apelimbart[apelimbart$Species=="Homo sapiens", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars],
                  fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 1", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]))
getstructure(list(fauxil[fauxil$Species=="Fauxil sp. 2", SSDvars]), forcematrix=TRUE)
}
\seealso{
\code{\link{SSDtest}}
}
